@@ -1,31 +1,22 @@
 <?php
 
+use App\Http\Controllers\AuthController;
 use App\Http\Controllers\CustomerController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
-/*
-|--------------------------------------------------------------------------
-| API Routes
-|--------------------------------------------------------------------------
-|
-| Here is where you can register API routes for your application. These
-| routes are loaded by the RouteServiceProvider within a group which
-| is assigned the "api" middleware group. Enjoy building your API!
-|
-*/
-//Route::post('/v1/customers', [CustomerController::class, 'store']);
+//public
+Route::get('customers', [CustomerController::class, 'index']);
+Route::post('register', [AuthController::class, 'register']);
+Route::post('login', [AuthController::class, 'login']);
 
-Route::post('customers', 'CustomerController@store');
-Route::get('customers', 'CustomerController@index');
-Route::put('customers/{id}', 'CustomerController@update')->where('id', '[0-9]+');
-Route::delete('customers/{id}', 'CustomerController@destroy')->where('id', '[0-9]+');
+//protected
+Route::group(['middleware' => ['auth:sanctum']], function () {
+    Route::post('customers', [CustomerController::class, 'store']);
+    Route::put('customers/{id}', [CustomerController::class, 'update']);
+    Route::delete('customers/{id}', [CustomerController::class, 'destroy']);
+    Route::post('logout', [AuthController::class, 'logout']);
 
-Route::group([
-    'prefix' => 'auth'
-], function () {
-    Route::post('login', 'AuthController@login');
-    Route::post('register', 'AuthController@register');
 });
 
 Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
